intent-mcp/src/requirements.txt
fastapi==0.116.1
fastmcp==2.11.3
httpx==0.28.1
mcp==1.13.0
openai==1.99.9
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2